+++ app/controllers/application_controller.rb
@@ -2,6 +2,24 @@ class ApplicationController < ActionController::Base
   before_action :authenticate_user!
   before_action :configure_permitted_parameters, if: :devise_controller?
 
+  before_action :authenticate_user!
+  after_action :verify_authorized, except: :index, unless: :skip_pundit?
+  after_action :verify_policy_scoped, only: :index, unless: :skip_pundit?
+  include Pundit
+  protect_from_forgery
+
+  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
+  def user_not_authorized
+    flash[:alert] = "You are not authorized to perform this action."
+    redirect_to(root_path)
+  end
+
+  private
+
+  def skip_pundit?
+    devise_controller? || params[:controller] =~ /(^(rails_)?admin)|(^pages$)/
+  end
+
   def configure_permitted_parameters
     # For additional fields in app/views/devise/registrations/new.html.erb
     # this allows the user to add a first name and last name on sign up
@@ -13,4 +31,5 @@ class ApplicationController < ActionController::Base
     # edit the array that contains :username with any other columns you want to add to users table i.e. :bio
     devise_parameter_sanitizer.permit(:account_update, keys: [:username])
   end
+
 end
