+++ app/controllers/dashboard_controller.rb
@@ -1,5 +1,9 @@
 class DashboardController < ApplicationController
+  # include Pundit
+  # before_action :verify_authorized, except: :show
+
   def show
     @user = current_user
+    authorize @user
   end
 end
