+++ app/policies/application_policy.rb
@@ -1,18 +1,6 @@
 class ApplicationPolicy
   attr_reader :user, :record
 
-  before_action :authenticate_user!
-  after_action :verify_authorized, except: :index, unless: :skip_pundit?
-  after_action :verify_policy_scoped, only: :index, unless: :skip_pundit?
-
-
-  # Uncomment when you *really understand* Pundit!
-  # rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
-  # def user_not_authorized
-  #   flash[:alert] = "You are not authorized to perform this action."
-  #   redirect_to(root_path)
-  # end
-
   def initialize(user, record)
     @user = user
     @record = record
@@ -59,10 +47,5 @@ class ApplicationPolicy
     end
   end
 
-  private
-
-  def skip_pundit?
-    devise_controller? || params[:controller] =~ /(^(rails_)?admin)|(^pages$)/
-  end
 
 end
