+++ app/policies/user_policy.rb
@@ -4,4 +4,10 @@ class UserPolicy < ApplicationPolicy
       scope.all
     end
   end
+
+  def create?
+    true
+  end
+
+
 end
